Patch to fix missing tile issue
