+++ src/components/Map.tsx
@@ -9,7 +9,7 @@ export interface IMapProps {
 }
 
 const Map = ({ center, zoom }: IMapProps) => (
-  <MapContainer center={center} zoom={zoom}>
+  <MapContainer style={{ height: '100vh' }} center={center} zoom={zoom}>
     <TileLayer
       attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
       url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
